Update README.md
PairWork

WePairHealth is community health worker placement and training platform which allows its members to search for a job, list their services, take work related courses, connect with organizations, events, lectures and tell their story about it all. WePairHealth is an MVP and due to the proprietary nature of the code I can only provide code snippets of my code. This repository contains SQL, .NET, and React code.
My role as a software engineer was to implement the functionality of the file uploader component and the share story components and functionalities. I was required to handle all the coding from the backend to the frontend. Below is a list of the actions I took to accomplish my responsibilities.

● Developed and structured the share story display, manage, and create functionalities from front-end to back-end.
● Tested and reworked the file uploader and it’s functionality from front-end to back-end.
● Created the necessary procedures in SQL utilizing joins, foreign keys and pagination.
● Built the API endpoints with .NET using C# and thoroughly tested with Postman.
● Designed components in React to maximize aesthetics and functionality.
● Utilized styling with CSS and Bootstrap.